% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{extra_rule_vars}
\alias{extra_rule_vars}
\title{Additional per-rule variables to display in error drill-down}
\usage{
extra_rule_vars(rule)
}
\description{
Additional per-rule variables to display in error drill-down
}
