% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{format_error_table}
\alias{format_error_table}
\title{Returns a DT::datatable object for displaying}
\usage{
format_error_table(
  error_table,
  value_table,
  rule = NULL,
  file = c("student", "course", "student_course", "graduation", "building", "room")
)
}
\arguments{
\item{error_table}{dataframe as returned in `errors` component of `get_stats_tables()`}

\item{file}{represents the file used to perform the audits i.e student, course, student_course, graduation, etc...}
}
\description{
Returns a DT::datatable object for displaying
}
