% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{format_rule_table}
\alias{format_rule_table}
\title{Returns a DT::datatable object for displaying}
\usage{
format_rule_table(rule_table)
}
\arguments{
\item{error_table}{dataframe as returned in `errors_byrule` component of `get_stats_tables()`}
}
\description{
Returns a DT::datatable object for displaying
}
