% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{format_summary_table}
\alias{format_summary_table}
\title{Produces a gt object for displaying the summary table (in summary and home tabs)}
\usage{
format_summary_table(summary_table)
}
\arguments{
\item{summary_table}{dataframe, as returned in `error_summary` component of
`get_stats_tables()` output}
}
\description{
Produces a gt object for displaying the summary table (in summary and home tabs)
}
