% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_app_data}
\alias{get_app_data}
\title{Returns a list of lists of tables used by the app. Top level is per UI tab,
each element as returned by `get_stats_tables()`}
\usage{
get_app_data(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{check_results}{Results of `utValidateR::do_checks()`, as returned by `load_data_in()`}
}
\description{
Returns a list of lists of tables used by the app. Top level is per UI tab,
each element as returned by `get_stats_tables()`
}
