% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_banner_table}
\alias{get_banner_table}
\title{Extract the table name from a banner field string}
\usage{
get_banner_table(banner_field)
}
\arguments{
\item{banner_field}{a string formatted like `"banner.<table>_<field>"`}
}
\description{
Extract the table name from a banner field string
}
