% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_data_values}
\alias{get_data_values}
\title{Returns a dataframe containing only columns of check_result_df that are data values (not status, etc)}
\usage{
get_data_values(check_result_df)
}
\description{
Returns a dataframe containing only columns of check_result_df that are data values (not status, etc)
}
