% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_pivot_id_cols}
\alias{get_pivot_id_cols}
\title{Returns the vector of id columns for the relevant file's error table}
\usage{
get_pivot_id_cols(
  file = c("student", "course", "student_course", "graduation", "building", "room")
)
}
\description{
This will need to be updated in future.
}
