% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{get_rule_info}
\alias{get_rule_info}
\title{returns a dataframe with columns <id_vars>, rule, description, <value vars>}
\usage{
get_rule_info(valuedf, rule, id_vars = c("row", "file"))
}
\description{
returns a dataframe with columns <id_vars>, rule, description, <value vars>
}
