% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_rule_variables}
\alias{get_rule_variables}
\title{Returns vector of variables referenced in the specified rule}
\usage{
get_rule_variables(rule, checklist = utValidateR::get_checklist())
}
\description{
Returns vector of variables referenced in the specified rule
}
