% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_stats_tables}
\alias{get_stats_tables}
\title{Returns a list of tables used by the app: `five_stats`, `error_summary`, and
(if `include_errors == TRUE`) `errors`}
\usage{
get_stats_tables(check_results, checklist, include_errors = TRUE)
}
\arguments{
\item{check_results}{a result of `do_checks()` or a list of such results}

\item{checklist}{the checklist used for check_results (the full checklist from utValidateR is fine)}

\item{include_errors}{if TRUE (default), include the large tibble enumerating all errors}
}
\description{
Returns a list of tables used by the app: `five_stats`, `error_summary`, and
(if `include_errors == TRUE`) `errors`
}
