% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_utValidateR_checklist}
\alias{get_utValidateR_checklist}
\title{Retrieves the checklist data object from the utValidateR package}
\usage{
get_utValidateR_checklist()
}
\description{
Assumes that the currently installed version of utValidateR was used to
generate the check_results object
}
