% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_in}
\alias{load_data_in}
\title{Pull in the results of utValidateR checks}
\usage{
load_data_in(session = shiny::getDefaultReactiveDomain())
}
\description{
Currenlty just loads saved dummy data from dev-data/
}
