% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_error_table}
\alias{make_error_table}
\title{This function will make the error table for each tab. This function should
take a user id and a perspective as input and should return a table. These
tables will contain all the data errors from utValidateR associated with the
perspective and the user. These should be obtainable from the utValidateR
functions.}
\usage{
make_error_table()
}
\description{
This function will make the error table for each tab. This function should
take a user id and a perspective as input and should return a table. These
tables will contain all the data errors from utValidateR associated with the
perspective and the user. These should be obtainable from the utValidateR
functions.
}
