% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_summary_table}
\alias{make_summary_table}
\title{Creates summary table for each tab.}
\usage{
make_summary_table(user_id, perspective)
}
\description{
This function should take a user id and a perspective as input and should
 return a table. These tables should be will be small, so I recommend we use
 the gt package. A description of the data points in the table is given in
 the home tab section. Each data point should be associated with a
 perspective. We may need to manipulate the output of the utValidateR package
 to get it.
}
