% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_result}
\alias{read_result}
\title{Reads the results of from either dummy data or production data}
\usage{
read_result(
  data_type = c("dummy", "edify"),
  file = c("student", "course", "student_course", "graduation", "building", "room")
)
}
\description{
Reads the results of from either dummy data or production data
}
