% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{title_logo}
\alias{title_logo}
\title{Create title logo}
\usage{
title_logo()
}
\description{
Returns UT logo with correct dimensions for app title.
#' Create title logo
}
\details{
Returns UT logo with correct dimensions for app title.
}
