% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{withModal}
\alias{withModal}
\title{Show a modal notification when evaluating an expression}
\usage{
withModal(expr, message, session)
}
\arguments{
\item{message}{Message to be displayed}

\item{session}{The session object passed to function given to shinyServer.}
}
\description{
Show a modal notification when evaluating an expression
}
